change time off job

+++ jobs/scheduling/TimeOffRequests.js
@@ -37,18 +37,11 @@ function handleTimeOffRequests() {
                 "start": request.start_time,
                 "end": request.end_time,
                 "user_id": request.user_id,
-                "location_id": global.config.locationID.regular_shifts,
+                "location_id": [global.config.locationID.regular_shifts, global.config.locationID.makeup_and_extra_shifts],
                 "unpublished": true
             };
 
             WhenIWork.get('shifts', shiftSearchParams, function(response) {
-
-                if (!response.shifts || response.shifts.length === 0) {
-                    console.log('No shifts found that fall within the range of timeoff request.');
-                    return false;
-                }
-
-                console.log('shift responses which fall within ', request.start_time, '**', request.end_time, ' &&& number of shifts returned for that time off request', response.shifts.length);
 
                 // Status code `2` represents approved requests.
                 var timeOffApprovalRequest = {
@@ -61,32 +54,35 @@ function handleTimeOffRequests() {
 
                 var batchPayload = [timeOffApprovalRequest];
 
-                //For each shift, create a delete request and an open shift to replace it
-                response.shifts.forEach(function(shift) {
-                    var shiftDeleteRequest = {
-                        "method": "delete",
-                        "url": "/2/shifts/" + shift.id,
-                        "params": {},
-                    };
+                if (!response.shifts || response.shifts.length === 0) {
+                    console.log('No shifts found that fall within the range of that timeoff request, but we\'re still approving that request.');
+                }
+                else {
+                    //For each shift, create a delete request and an open shift to replace it
+                    response.shifts.forEach(function(shift) {
+                        var shiftDeleteRequest = {
+                            "method": "delete",
+                            "url": "/2/shifts/" + shift.id,
+                            "params": {},
+                        };
 
-                    batchPayload.push(shiftDeleteRequest);
+                        batchPayload.push(shiftDeleteRequest);
 
-                    var newOpenShiftRequest = {
-                        "method": "post",
-                        "url": "/2/shifts",
-                        "params": {
-                            "start_time": shift.start_time,
-                            "end_time": shift.end_time,
-                            "notes": "SHIFT COVERAGE",
-                            "published": true,
-                            "location_id": global.config.locationID
-                                                 .makeup_and_extra_shifts,
-                        }
-                    };
-
-                    batchPayload.push(newOpenShiftRequest);
-                });
-
+                        var newOpenShiftRequest = {
+                            "method": "post",
+                            "url": "/2/shifts",
+                            "params": {
+                                "start_time": shift.start_time,
+                                "end_time": shift.end_time,
+                                "notes": "SHIFT COVERAGE",
+                                "published": true,
+                                "location_id": global.config.locationID
+                                                     .makeup_and_extra_shifts,
+                            }
+                        };
+                        batchPayload.push(newOpenShiftRequest);
+                    });
+                }
                 //Send the time off request approval and all shift deletions/open shift creations to batch
                 WhenIWork.post('batch', batchPayload, function(response) {
                     console.log(response);
